Fixed training and evaluation bugs
- Fixed bug where entities got removed from trainings set, fixes #43
- Fixed bug where encoded triples consisted of string integers and caused
a conversion error, fixes #44
- Changed output of nodes from train_val_nodes.csv to train_nodes.csv and
val_nodes.csv
- Fixed Evaluation GUI to include negative sets

+++ src/openbiolink/utils.py
@@ -496,7 +496,7 @@ def _get_corrupted_examples(corrupted_triples, pos_examples, filtered):
     else:
         corrupted_triples_df[globConst.VALUE_COL_NAME] = 0
         # corrupted_triples[globConst.VALUE_COL_NAME][neg_indices] = 0
-    return corrupted_triples_df.values
+    return corrupted_triples_df.astype(float).astype(int).values
 
 
 def _group_corrupted_examples(corrupted_dict, col_names):
